+++ wordless/wl_nlp/wl_pos_tagging.py
@@ -71,6 +71,7 @@ def wl_pos_tag(main, inputs, lang, pos_tagger = 'default', tagset = 'default', s
     else:
         texts_tagged = []
         tags = []
+        convert_srp_script = False
 
         settings = main.settings_custom['pos_tagging']
 
@@ -82,6 +83,11 @@ def wl_pos_tag(main, inputs, lang, pos_tagger = 'default', tagset = 'default', s
 
         if separator == 'default':
             separator = main.settings_custom['pos_tagging']['pos_tagger_settings']['separator_between_tokens_pos_tags']
+
+        # Modify the language after the default POS tagger is loaded and before the POS tagger is initialized
+        if lang == 'srp_cyrl' and pos_tagger == 'stanza_srp_latn':
+            lang = 'srp_latn'
+            convert_srp_script = True
 
         wl_nlp_utils.init_word_tokenizers(
             main,
@@ -97,6 +103,9 @@ def wl_pos_tag(main, inputs, lang, pos_tagger = 'default', tagset = 'default', s
         tags_universal = []
 
         if isinstance(inputs, str):
+            if convert_srp_script:
+                inputs = wl_nlp_utils.to_srp_latn((inputs,))[0]
+
             # spaCy and modern-botok
             if pos_tagger.startswith('spacy_') or pos_tagger == 'modern_botok_bod':
                 lang_spacy = wl_conversion.remove_lang_code_suffixes(lang)
@@ -159,6 +168,9 @@ def wl_pos_tag(main, inputs, lang, pos_tagger = 'default', tagset = 'default', s
                     tags.extend(tags_line)
         else:
             texts, token_properties = wl_texts.split_texts_properties(inputs)
+
+            if convert_srp_script:
+                texts = wl_nlp_utils.to_srp_latn(texts)
 
             # spaCy and modern-botok
             if pos_tagger.startswith('spacy_') or pos_tagger == 'modern_botok_bod':
@@ -287,6 +299,9 @@ def wl_pos_tag(main, inputs, lang, pos_tagger = 'default', tagset = 'default', s
         tags = [f'{separator}{tag}' for tag in tags]
 
         if isinstance(inputs, str):
+            if convert_srp_script:
+                texts_tagged = wl_nlp_utils.to_srp_cyrl(texts_tagged)
+
             return wl_texts.to_tokens(
                 texts_tagged,
                 lang = lang,